Add Yearly recurrence UI.
Refactor recurrence setup and update calendar item times

Refactored recurrence setup logic by moving it to a new module `add-edit-recurrence.js` and simplifying `add-edit-event.ts` by
removing redundant functions.

Updated `CalendarController.cs` to set `StartTime` and `EndTime` for `model.CalendarItem` using the user's timezone and defaulting the end time to 10 minutes after the start time.

Improved month selection and recurrence rule handling in `CalendarItemViewModel.cs`.

Updated HTML and CSS for yearly repeat options, corrected a typo in class names, and ensured start and end times are populated from model data.
